backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
flask-migrate==4.0.7
flask-cors==4.0.1
python-dotenv==1.0.1
marshmallow==3.21.3
sqlalchemy==2.0.36
PyMySQL==1.1.1
cryptography==42.0.8
werkzeug==3.0.6